Use correct width for codeblocks
The previous settings for the listings package use \textwidth as line width
but \textwidth is "The full horizontal width of the entire page body."
which is too large in (nested) lists. Therefore use \linewidth which is
always the width of the current environment for the listings
environment.

Tested by generating the German manual.

Fixes siduction/sidu-manual#44